refactor: move email parse + IMAP fetch into the framework (stack.*)
Email is an ingestion channel, so its primitives belong in stack.* where
every surface imports them: the mail bot (core), the archivist (docs), the
host CLI, and tests. Moves ParsedEmail/parse_email to stack.email_message and
the IMAP MailFetcher to stack.mail_fetcher; extractors re-exports the parser
for back-compat. No behaviour change.
